Double FIR coefficient bitwidth.
* Improve dynamic range of FIR filter by doubling the coefficient bitwidth.
* New Ram2 module.
* Trickle change up to library and daemon.

+++ driver/whitebox_ioctl.h
@@ -44,7 +44,7 @@ typedef struct whitebox_args {
         struct {
             uint8_t bank;
             uint8_t n;
-            int16_t coeff[WF_COEFF_COUNT];
+            int32_t coeff[WF_COEFF_COUNT];
         } fir;
     } flags;
     uint8_t locked;
